sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "untrace-sdk"
version = "0.1.2"
description = "LLM observability SDK for Python"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Untrace", email = "hello@untrace.dev"},
]
keywords = ["llm", "observability", "tracing", "monitoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-semantic-conventions>=0.40b0",
    "opentelemetry-instrumentation>=0.40b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]
test = [
    "openai>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/untrace-dev/untrace-sdk"
Repository = "https://github.com/untrace-dev/untrace-sdk"
Documentation = "https://docs.untrace.dev"
"Bug Tracker" = "https://github.com/untrace-dev/untrace-sdk/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/untrace"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["untrace"]

[dependency-groups]
dev = [
    "black>=24.8.0",
    "build>=1.2.2.post1",
    "isort>=5.13.2",
    "mypy>=1.14.1",
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.5",
    "ruff>=0.12.12",
    "twine>=4.0.0",
    "pytest-cov>=5.0.0",
]
